Create The Number Game
The approach is to generate a random number using Math. random() method in Java. Now using a loop, take K input from the user, and for each input print whether the number is smaller or larger than the actual number. If within K trials the user guessed the number correctly, print that the user won.